requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
sentence-transformers==2.3.1
qdrant-client==1.7.0
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10